ShipEF: A package containing functions to assign marine vessel emission factors

The ShipEF package by the U.S. Environmental Protection Agency provides functions for assigning emission factors to commercial marine vessels (CMV). For various pollutants, emission factors depend on vessel characteristics, vessel operating modes, and/or vessel location. This package can be used in conjunction with ship power estimates (see the ShipPowerModel package) and ship activity data to estimate CMV air emissions.

Primary ShipEF functions

Each pollutant included in this package has its own function to determine the appropriate emissions factor. There is also a low load adjustment factor, which applies to the emission factors for each pollutant in certain operating conditions.

calcEF_CO - Carbon monoxide emissions factor

calcEF_CO2 - Carbon dioxide emissions factor

calcEF_HC - Hydrocarbons emissions factor

calcEF_NOx - Oxides of nitrogen emissions factor

calcEF_PM - Particulate matter emissions factor

calcEF_SO2 - Sulfur dioxide emissions factor

calcLLAF - Low load adjustment factor

calcEF - Calculates emission factors for all pollutants and applies the LLAF

Vignettes

An example of how to use this package can be found in the following vignette: vignette("OverviewOfShipEF", package="ShipEF")